# swagger.yaml
openapi: 3.0.0
info:
  title: API Port de Plaisance
  version: 1.0.0
  description: Documentation de l'API
paths:
...
  /catways:
    get:
      summary: Récupère la liste des catways
      tags: [Catways]
      responses:
        200: { description: Liste récupérée }
        501: { description: Erreur serveur }
  /catways/{id}:
    post:
      summary: Récupère un catway
      tags: [Catways]
      responses:
        200: { description: Trouvé }
        404: { description: Non trouvé }
    patch:
      summary: Met à jour un catway
      tags: [Catways]
      responses:
        200: { description: Mis à jour }
        404: { description: Non trouvé }
    delete:
      summary: Supprime un catway
      tags: [Catways]
      responses:
        200: { description: Supprimé }
...